+++ src/hooks/useItems.ts
@@ -1,0 +1,14 @@
+import { useState } from "react";
+import { generateItems } from "../utils";
+
+export const useItems = () => {
+  const [items, setItems] = useState(generateItems(1000));
+
+  const addItems = () => {
+    setItems((prevItems) => [
+      ...prevItems,
+      ...generateItems(1000, prevItems.length),
+    ]);
+  };
+  return { items, addItems };
+};
